refactor: タスクの追加処理が肥大化したため
タスクの追加フィールドと追加処理を別ファイルに分離
