+++ todo-app/src/App.tsx
@@ -1,14 +1,13 @@
-import { CheckIcon } from "@chakra-ui/icons";
-import { Input, InputGroup, InputRightElement } from "@chakra-ui/input";
 import { Container, Text } from "@chakra-ui/layout";
 import { Center } from "@chakra-ui/layout";
-import { useRef, useState } from "react";
+import { useState } from "react";
+import InputNewTaskField from "./InputNewTaskField";
+
+export interface Task {
+  text: string;
+}
 
 const App = () => {
-  interface Task {
-    text: string;
-  }
-
   const [tasks, setTasks] = useState<Task[]>([
     { text: "Task1" },
     { text: "Task2" },
@@ -18,27 +17,6 @@ const App = () => {
     },
   ]);
 
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const validInputText = (): boolean => {
-    return inputRef.current?.value !== "";
-  };
-
-  const addTask = (text: string) => {
-    const newTask = { text: text };
-    const newTasks = [...tasks, newTask];
-    setTasks(newTasks);
-  };
-
-  const resetField = () => {};
-
-  const onSubmitTask = () => {
-    if (validInputText()) {
-      addTask(inputRef.current?.value!);
-      resetField();
-    }
-  };
-
   return (
     <Container maxW="xl" centerContent>
       {tasks.map((task: Task) => (
@@ -46,23 +24,7 @@ const App = () => {
           <Text isTruncated>{task.text}</Text>
         </Center>
       ))}
-      <InputGroup m={1}>
-        <Input
-          borderRadius="0"
-          placeholder="Input new task"
-          ref={inputRef}
-          onKeyPress={(e) => {
-            if (e.key === "Enter") {
-              onSubmitTask();
-            }
-          }}
-        />
-        <InputRightElement
-          children={<CheckIcon color="green.500" />}
-          _hover={{ bg: "green.100", borderRadius: "50%" }}
-          onClick={onSubmitTask}
-        />
-      </InputGroup>
+      <InputNewTaskField tasks={tasks} setTasks={setTasks} />
     </Container>
   );
 };
